requirements.txt
praw==7.8.1
python-dotenv==1.0.0
#openai==0.28.1
openai==1.58.1